Rename RGBELoader to HDRLoader across examples/jsm, update imports/usages/docs, and adjust HDRCubeTextureLoader to depend on HDRLoader
