+++ docs/api/ar/loaders/DataTextureLoader.html
@@ -20,7 +20,7 @@
 		<h2>أمثلة (Examples)</h2>
 		<p>
 		انظر
-		[link:https://github.com/mrdoob/three.js/blob/master/examples/jsm/loaders/RGBELoader.js RGBELoader]
+		[link:https://github.com/mrdoob/three.js/blob/master/examples/jsm/loaders/HDRLoader.js HDRLoader]
 		لمثال على فئة مشتقة.
 		</p>
