+++ docs/api/en/loaders/DataTextureLoader.html
@@ -20,7 +20,7 @@
 		<h2>Examples</h2>
 		<p>
 			See the
-			[link:https://github.com/mrdoob/three.js/blob/master/examples/jsm/loaders/RGBELoader.js RGBELoader] 
+			[link:https://github.com/mrdoob/three.js/blob/master/examples/jsm/loaders/HDRLoader.js HDRLoader] 
 			for an example of a derived class.
 		</p>
