+++ docs/api/it/loaders/DataTextureLoader.html
@@ -19,7 +19,7 @@
 		<h2>Esempi</h2>
 
 		<p>
-			Vedi l'[link:https://github.com/mrdoob/three.js/blob/master/examples/jsm/loaders/RGBELoader.js RGBELoader]
+			Vedi l'[link:https://github.com/mrdoob/three.js/blob/master/examples/jsm/loaders/HDRLoader.js HDRLoader]
       per un esempio di una classe derivata.
 		</p>
