+++ docs/api/zh/loaders/DataTextureLoader.html
@@ -20,7 +20,7 @@
 		<h2>例子</h2>
 
 		<p>
-			请参考[link:https://github.com/mrdoob/three.js/blob/master/examples/jsm/loaders/RGBELoader.js RGBELoader]
+			请参考[link:https://github.com/mrdoob/three.js/blob/master/examples/jsm/loaders/HDRLoader.js HDRLoader]
 			这个子类的例子。
 		</p>
