+++ examples/jsm/Addons.js
@@ -108,7 +108,7 @@ export * from './loaders/PCDLoader.js';
 export * from './loaders/PDBLoader.js';
 export * from './loaders/PLYLoader.js';
 export * from './loaders/PVRLoader.js';
-export * from './loaders/RGBELoader.js';
+export * from './loaders/HDRLoader.js';
 export * from './loaders/UltraHDRLoader.js';
 export * from './loaders/STLLoader.js';
 export * from './loaders/SVGLoader.js';
