+++ examples/misc_exporter_ktx2.html
@@ -26,7 +26,7 @@
 
 			import { OrbitControls } from 'three/addons/controls/OrbitControls.js';
 			import { KTX2Exporter } from 'three/addons/exporters/KTX2Exporter.js';
-			import { RGBELoader } from 'three/addons/loaders/RGBELoader.js';
+			import { HDRLoader } from 'three/addons/loaders/HDRLoader.js';
 			import { GUI } from 'three/addons/libs/lil-gui.module.min.js';
 
 			let scene, camera, renderer, exporter, mesh, controls, renderTarget, dataTexture;
@@ -55,7 +55,7 @@
 				scene = new THREE.Scene();
 
 				exporter = new KTX2Exporter();
-				const rgbeloader = new RGBELoader();
+				const rgbeloader = new HDRLoader();
 
 				//
