+++ examples/webgl_loader_gltf_anisotropy.html
@@ -29,7 +29,7 @@
 
 			import { OrbitControls } from 'three/addons/controls/OrbitControls.js';
 			import { GLTFLoader } from 'three/addons/loaders/GLTFLoader.js';
-			import { RGBELoader } from 'three/addons/loaders/RGBELoader.js';
+			import { HDRLoader } from 'three/addons/loaders/HDRLoader.js';
 
 			let renderer, scene, camera, controls;
 
@@ -56,7 +56,7 @@
 				controls.addEventListener( 'change', render );
 				controls.update();
 
-				const rgbeLoader = new RGBELoader().setPath( 'textures/equirectangular/' );
+				const rgbeLoader = new HDRLoader().setPath( 'textures/equirectangular/' );
 				const gltfLoader = new GLTFLoader().setPath( 'models/gltf/' );
 
 				const [ texture, gltf ] = await Promise.all( [
