+++ examples/webgl_loader_gltf_iridescence.html
@@ -29,7 +29,7 @@
 
 			import { OrbitControls } from 'three/addons/controls/OrbitControls.js';
 			import { GLTFLoader } from 'three/addons/loaders/GLTFLoader.js';
-			import { RGBELoader } from 'three/addons/loaders/RGBELoader.js';
+			import { HDRLoader } from 'three/addons/loaders/HDRLoader.js';
 
 			let renderer, scene, camera, controls;
 
@@ -59,7 +59,7 @@
 				controls.target.set( 0, 0.2, 0 );
 				controls.update();
 
-				const rgbeLoader = new RGBELoader()
+				const rgbeLoader = new HDRLoader()
 					.setPath( 'textures/equirectangular/' );
 
 				const gltfLoader = new GLTFLoader().setPath( 'models/gltf/' );
