+++ examples/webgl_loader_texture_hdr.html
@@ -28,7 +28,7 @@
 
 			import { GUI } from 'three/addons/libs/lil-gui.module.min.js';
 
-			import { RGBELoader } from 'three/addons/loaders/RGBELoader.js';
+			import { HDRLoader } from 'three/addons/loaders/HDRLoader.js';
 
 			const params = {
 				exposure: 2.0
@@ -55,7 +55,7 @@
 
 				camera = new THREE.OrthographicCamera( - aspect, aspect, 1, - 1, 0, 1 );
 
-				new RGBELoader()
+				new HDRLoader()
 					.load( 'textures/memorial.hdr', function ( texture, textureData ) {
 
 						//console.log( textureData );
