+++ examples/webgl_loader_usdz.html
@@ -34,7 +34,7 @@
 			import * as THREE from 'three';
 
 			import { OrbitControls } from 'three/addons/controls/OrbitControls.js';
-			import { RGBELoader } from 'three/addons/loaders/RGBELoader.js';
+			import { HDRLoader } from 'three/addons/loaders/HDRLoader.js';
 			import { USDLoader } from 'three/addons/loaders/USDLoader.js';
 
 			let camera, scene, renderer;
@@ -48,7 +48,7 @@
 
 				scene = new THREE.Scene();
 
-				const rgbeLoader = new RGBELoader()
+				const rgbeLoader = new HDRLoader()
 					.setPath( 'textures/equirectangular/' );
 
 				const usdzLoader = new USDLoader()
