+++ examples/webgl_materials_envmaps_groundprojected.html
@@ -36,7 +36,7 @@
 			import { GroundedSkybox } from 'three/addons/objects/GroundedSkybox.js';
 			import { GLTFLoader } from 'three/addons/loaders/GLTFLoader.js';
 			import { DRACOLoader } from 'three/addons/loaders/DRACOLoader.js';
-			import { RGBELoader } from 'three/addons/loaders/RGBELoader.js';
+			import { HDRLoader } from 'three/addons/loaders/HDRLoader.js';
 
 			const params = {
 				height: 15,
@@ -61,7 +61,7 @@
 
 				scene = new THREE.Scene();
 
-				const hdrLoader = new RGBELoader();
+				const hdrLoader = new HDRLoader();
 				const envMap = await hdrLoader.loadAsync( 'textures/equirectangular/blouberg_sunrise_2_1k.hdr' );
 				envMap.mapping = THREE.EquirectangularReflectionMapping;
