+++ examples/webgl_materials_physical_transmission.html
@@ -26,7 +26,7 @@
 
 			import { GUI } from 'three/addons/libs/lil-gui.module.min.js';
 			import { OrbitControls } from 'three/addons/controls/OrbitControls.js';
-			import { RGBELoader } from 'three/addons/loaders/RGBELoader.js';
+			import { HDRLoader } from 'three/addons/loaders/HDRLoader.js';
 
 			const params = {
 				color: 0xffffff,
@@ -48,7 +48,7 @@
 
 			let mesh;
 
-			const hdrEquirect = new RGBELoader()
+			const hdrEquirect = new HDRLoader()
 				.setPath( 'textures/equirectangular/' )
 				.load( 'royal_esplanade_1k.hdr', function () {
