+++ examples/webgl_materials_physical_transmission_alpha.html
@@ -65,7 +65,7 @@
 
 			import { GUI } from 'three/addons/libs/lil-gui.module.min.js';
 			import { OrbitControls } from 'three/addons/controls/OrbitControls.js';
-			import { RGBELoader } from 'three/addons/loaders/RGBELoader.js';
+			import { HDRLoader } from 'three/addons/loaders/HDRLoader.js';
 			import { GLTFLoader } from 'three/addons/loaders/GLTFLoader.js';
 
 			const params = {
@@ -89,7 +89,7 @@
 
 			let mesh, material;
 
-			const hdrEquirect = new RGBELoader()
+			const hdrEquirect = new HDRLoader()
 				.setPath( 'textures/equirectangular/' )
 				.load( 'royal_esplanade_1k.hdr', function () {
