+++ examples/webgl_postprocessing_material_ao.html
@@ -36,7 +36,7 @@
 			import Stats from 'three/addons/libs/stats.module.js';
 			import { GUI } from 'three/addons/libs/lil-gui.module.min.js';
 			import { OrbitControls } from 'three/addons/controls/OrbitControls.js';
-			import { RGBELoader } from 'three/addons/loaders/RGBELoader.js';
+			import { HDRLoader } from 'three/addons/loaders/HDRLoader.js';
 			import { PLYLoader } from 'three/addons/loaders/PLYLoader.js';
 			import { EffectComposer } from 'three/addons/postprocessing/EffectComposer.js';
 			import { RenderPass } from 'three/addons/postprocessing/RenderPass.js';
@@ -78,7 +78,7 @@
 				renderer.shadowMap.enabled = sceneParameters.shadow;
 
 				const plyLoader = new PLYLoader();
-				const rgbeloader = new RGBELoader();
+				const rgbeloader = new HDRLoader();
 
 				camera = new THREE.PerspectiveCamera( 40, window.innerWidth / window.innerHeight, 1, 50 );
 				camera.position.set( 0, 3, 5 );
