+++ examples/webgpu_compute_cloth.html
@@ -31,7 +31,7 @@
 
 			import { GUI } from 'three/addons/libs/lil-gui.module.min.js';
 			import { OrbitControls } from 'three/addons/controls/OrbitControls.js';
-			import { RGBELoader } from 'three/addons/loaders/RGBELoader.js';
+			import { HDRLoader } from 'three/addons/loaders/HDRLoader.js';
 
 			let renderer, scene, camera, controls;
 
@@ -89,7 +89,7 @@
 				controls.target.set( 0, - 0.1, 0 );
 				controls.update();
 
-				const rgbeLoader = new RGBELoader().setPath( 'textures/equirectangular/' );
+				const rgbeLoader = new HDRLoader().setPath( 'textures/equirectangular/' );
 
 				const hdrTexture = await rgbeLoader.loadAsync( 'royal_esplanade_1k.hdr' );
 				hdrTexture.mapping = THREE.EquirectangularReflectionMapping;
