+++ examples/webgpu_compute_particles_fluid.html
@@ -32,7 +32,7 @@
 
 			import { GUI } from 'three/addons/libs/lil-gui.module.min.js';
 			import { OrbitControls } from 'three/addons/controls/OrbitControls.js';
-			import { RGBELoader } from 'three/addons/loaders/RGBELoader.js';
+			import { HDRLoader } from 'three/addons/loaders/HDRLoader.js';
 			import * as BufferGeometryUtils from 'three/addons/utils/BufferGeometryUtils.js';
 			import WebGPU from 'three/addons/capabilities/WebGPU.js';
 
@@ -89,7 +89,7 @@
 				controls.maxPolarAngle = Math.PI * 0.35;
 				controls.touches = { TWO: THREE.TOUCH.DOLLY_ROTATE };
 
-				const rgbeLoader = new RGBELoader().setPath( 'textures/equirectangular/' );
+				const rgbeLoader = new HDRLoader().setPath( 'textures/equirectangular/' );
 
 				const hdrTexture = await rgbeLoader.loadAsync( 'royal_esplanade_1k.hdr' );
 				hdrTexture.mapping = THREE.EquirectangularReflectionMapping;
