+++ examples/webgpu_materials_transmission.html
@@ -28,7 +28,7 @@
 
 			import { GUI } from 'three/addons/libs/lil-gui.module.min.js';
 			import { OrbitControls } from 'three/addons/controls/OrbitControls.js';
-			import { RGBELoader } from 'three/addons/loaders/RGBELoader.js';
+			import { HDRLoader } from 'three/addons/loaders/HDRLoader.js';
 
 			const params = {
 				color: 0xffffff,
@@ -49,7 +49,7 @@
 
 			let mesh;
 
-			const hdrEquirect = new RGBELoader()
+			const hdrEquirect = new HDRLoader()
 				.setPath( 'textures/equirectangular/' )
 				.load( 'royal_esplanade_1k.hdr', function () {
