+++ examples/webgpu_tsl_procedural_terrain.html
@@ -32,7 +32,7 @@
 
 			import { GUI } from 'three/addons/libs/lil-gui.module.min.js';
 			import { OrbitControls } from 'three/addons/controls/OrbitControls.js';
-			import { RGBELoader } from 'three/addons/loaders/RGBELoader.js';
+			import { HDRLoader } from 'three/addons/loaders/HDRLoader.js';
 
 			let camera, scene, renderer, controls, drag;
 
@@ -50,7 +50,7 @@
 
 				// environment
 
-				const rgbeLoader = new RGBELoader();
+				const rgbeLoader = new HDRLoader();
 				rgbeLoader.load( './textures/equirectangular/pedestrian_overpass_1k.hdr', ( environmentMap ) => {
 
 					environmentMap.mapping = THREE.EquirectangularReflectionMapping;
