+++ examples/webxr_ar_lighting.html
@@ -24,7 +24,7 @@
 		<script type="module">
 
 			import * as THREE from 'three';
-			import { RGBELoader } from 'three/addons/loaders/RGBELoader.js';
+			import { HDRLoader } from 'three/addons/loaders/HDRLoader.js';
 			import { ARButton } from 'three/addons/webxr/ARButton.js';
 			import { XREstimatedLight } from 'three/addons/webxr/XREstimatedLight.js';
 
@@ -89,7 +89,7 @@
 
 				//
 
-				new RGBELoader()
+				new HDRLoader()
 					.setPath( 'textures/equirectangular/' )
 					.load( 'royal_esplanade_1k.hdr', function ( texture ) {
